UHM-5998 - Remove ci and test-resources maven submodules, convert 2 IT tests to normal tests (these are already ignored), move resources over to pihcore module.
