+++ api/src/test/resources/fromMirebalaisMetadataModule.xml
@@ -1,0 +1,18 @@
+<dataset>
+    <patient_identifier_type patient_identifier_type_id="100" name="ZL EMR ID"
+                             description="A unique identifier issued to all patients by the ZL EMR.  Blocks of this identifier are issued to each site to prevent duplication, so the identifier is unique across all sites.  The identifier uses six digits and is alphanumeric base 30, omitting the letters B, I, O, Q, S and Z as these can be confused with 8, 1, 0, 0, 5 and 2."
+                             format="" creator="1" date_created="2005-09-22 00:00:00.0"
+                             required="false"
+                             validator="org.openmrs.module.idgen.validator.LuhnMod30IdentifierValidator" retired="false"
+                             uuid="a541af1e-105c-40bf-b345-ba1fd6a59b85"/>
+    <patient_identifier_type patient_identifier_type_id="101" name="Nimewo Dosye" description="Patient's Dossier number"
+                             format="" creator="1" date_created="2005-09-22 00:00:00.0"
+                             required="false"
+                             validator="org.openmrs.module.idgen.validator.LuhnMod30IdentifierValidator" retired="false"
+                             uuid="e66645eb-03a8-4991-b4ce-e87318e37566"/>
+    <patient_identifier_type patient_identifier_type_id="102" name="External Nimewo Dosye" description="External Patient's Dossier number"
+                             format="" creator="1" date_created="2005-09-22 00:00:00.0"
+                             required="false"
+                             validator="org.openmrs.module.idgen.validator.LuhnMod30IdentifierValidator" retired="false"
+                             uuid="9dbea4d4-35a9-4793-959e-952f2a9f5347"/>
+</dataset>
