+++ api/src/test/resources/test-hibernate.cfg.xml
@@ -34,8 +34,13 @@
         <mapping resource="ImportedPackage.hbm.xml"/>
         <mapping resource="ExportedPackage.hbm.xml"/>
 
+            <mapping resource="OutboundQueue.hbm.xml" />
         <mapping resource="NamePhonetic.hbm.xml" />
+
         <mapping resource="ReportDesign.hbm.xml" />
+            <mapping resource="ReportRequest.hbm.xml"/>
+
+            <mapping resource="AppFramework.hbm.xml"/>
 
         <mapping resource="MetadataSource.hbm.xml" />
         <mapping resource="MetadataTermMapping.hbm.xml" />
